Slim down the list of installed plugins
I have tried to reduce the number of plugins installed by default,
based on my experience and what was already there.

Also, rather than just dumping the list of all plugins (including deps),
I just list the plugins that I intended to include. I think that makes
the intentions clearer.

Obviously, that has been tested, and Jenkins install all the deps on startup.

Result: I list only 5 plugins - Jenkins installs a total of 50 deps on top of that!
That's an improvement (albeit not major) compared to the 70 we had.