Use a static Jenkisn configuration file when building the container
Before making any particular customisation, we start committing a "vanilla"
config.xml based on the current status of the Jenkins installation.

At some point, we can become more clever and create a template for it, and
pass values, such as the number of executors, ... However, let's keep it
simple and move fast for now.

Now that we use the config.xml file, we don't need to specify the number
of executors as we were doing, thus I removed the file from Dockerfile.